+++ Modules/Intent.Modules.CSharp.Importer/Importer/StereotypeExtensions.cs
@@ -1,62 +1,35 @@
+using Intent.Persistence;
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using Intent.IArchitect.Agent.Persistence.Model;
-using Intent.IArchitect.Agent.Persistence.Model.Common;
 
 namespace Intent.Modules.CSharp.Importer.Importer;
 
 internal static class StereotypeExtensions
 {
-    public static StereotypePersistable GetOrCreateStereotype(
-        this ElementPersistable element,
+    public static IStereotypePersistable GetOrCreateStereotype(
+        this IElementPersistable element,
         string stereotypeDefinitionId,
-        Action<StereotypePersistable>? initAction = null)
+        string name, 
+        string packageDefinitionId, 
+        string packageDefinitionName,
+        Action<IStereotypePersistable>? initAction = null)
     {
-        element.Stereotypes ??= new List<StereotypePersistable>();
-
         var stereotype = element.Stereotypes.SingleOrDefault(x => x.DefinitionId == stereotypeDefinitionId);
-        if (stereotype == null)
-        {
-            if (initAction == null)
-            {
-                throw new InvalidOperationException(
-                    $"Stereotype Definition Id [{stereotypeDefinitionId}] cannot be found and cannot be created without {nameof(initAction)} being specified");
-            }
-            stereotype = new StereotypePersistable
-            {
-                DefinitionId = stereotypeDefinitionId
-            };
-            initAction(stereotype);
-            element.Stereotypes.Add(stereotype);
-        }
+        stereotype ??= element.Stereotypes.Add(stereotypeDefinitionId, name, packageDefinitionId, packageDefinitionName, initAction);
 
         return stereotype;
     }
 
-    public static StereotypePropertyPersistable GetOrCreateProperty(
-        this StereotypePersistable stereotype,
+    public static IStereotypePropertyPersistable GetOrCreateProperty(
+        this IStereotypePersistable stereotype,
         string propertyId,
-        Action<StereotypePropertyPersistable>? initAction = null)
+        string name,
+        string value,
+        Action<IStereotypePropertyPersistable>? initAction = null)
     {
-        stereotype.Properties ??= new List<StereotypePropertyPersistable>();
-
         var property = stereotype.Properties.SingleOrDefault(p => p.DefinitionId == propertyId);
-        if (property == null)
-        {
-            if (initAction == null)
-            {
-                throw new InvalidOperationException(
-                    $"Stereotype Definition Id [{stereotype.DefinitionId}] cannot create a property with Id [{propertyId}] without {nameof(initAction)} being specified");
-            }
-            property = new StereotypePropertyPersistable
-            {
-                DefinitionId = propertyId,
-                IsActive = true
-            };
-            initAction(property);
-            stereotype.Properties.Add(property);
-        }
+        property ??= stereotype.Properties.Add(propertyId, name, value, initAction);
 
         return property;
     }