+++ Modules/Intent.Modules.CSharp.Importer/release-notes.md
@@ -4,6 +4,7 @@
 - Fixed: Return types correct set on service operations when importing from C# files.
 - Fixed: Software Factory will no longer crash after importing C# files without first saving the designer.
 - Fixed: Internal refactor to ensure that `Intent.Modules.Importer.FileDirectoryPreviewTask` no longer exists as other importers who have shared code could end up registering the same Task ID and cause a crash.
+- Fixed: Removed hard dependency on ElementPersistable when creating the CSharp stereotype.
 
 ### Version 1.0.4